requirements.txt
Django==1.6
django-mptt 0.6.0
